requirements.txt
pandas
numpy
scikit-learn
joblib
spacy
nltk
rapidfuzz
pymupdf
python-docx
streamlit
